Fixed Daily time series
when selecting #2 option which was daily it was not working after talking with Cole we found that daily was using a premium selection to which we thought changing might fix along with another "optional" change according to the API website which was output size. It now works and prints a graph.